rename to useFunctionWatcher
